Complete existing index creation (1:N relation reference)
The index creation logic was missing the creation of a databas index based on columns in a child table that refer to a parent table. The so called foreign key contraint is not enforced in the database. Yet the relevant columns are used in the API for retreiving the data. Also it can be used by database users who query on the tables directly.